HUM-850: some client interface cleanup
Break up directclient.go.
Clean up interfaces.
Get rid of unused classes
Make it so directClient isn't just dead code.
