+++ proxyserver/middleware/context.go
@@ -90,16 +90,16 @@ type AuthorizeFunc func(r *http.Request) (bool, int)
 type subrequestCopy func(dst, src *http.Request)
 
 type ProxyContextMiddleware struct {
-	next              http.Handler
-	log               srv.LowLevelLogger
-	Cache             ring.MemcacheRing
-	proxyDirectClient *client.ProxyDirectClient
-	debugResponses    bool
+	next               http.Handler
+	log                srv.LowLevelLogger
+	Cache              ring.MemcacheRing
+	proxyClientFactory client.ProxyClient
+	debugResponses     bool
 }
 
 type ProxyContext struct {
 	*ProxyContextMiddleware
-	C                client.ProxyClient
+	C                client.RequestClient
 	Authorize        AuthorizeFunc
 	RemoteUsers      []string
 	StorageOwner     bool
@@ -328,7 +328,7 @@ func (m *ProxyContextMiddleware) ServeHTTP(writer http.ResponseWriter, request *
 		TxId:                   transId,
 		status:                 500,
 		accountInfoCache:       make(map[string]*AccountInfo),
-		C:                      client.NewProxyClient(m.proxyDirectClient, m.Cache, make(map[string]*client.ContainerInfo), logr),
+		C:                      m.proxyClientFactory.NewRequestClient(m.Cache, make(map[string]*client.ContainerInfo), logr),
 	}
 	// we'll almost certainly need the AccountInfo and ContainerInfo for the current path, so pre-fetch them in parallel.
 	apiRequest, account, container, _ := getPathParts(request)
@@ -381,14 +381,14 @@ func (m *ProxyContextMiddleware) ServeHTTP(writer http.ResponseWriter, request *
 	m.next.ServeHTTP(newWriter, request)
 }
 
-func NewContext(debugResponses bool, mc ring.MemcacheRing, log srv.LowLevelLogger, proxyDirectClient *client.ProxyDirectClient) func(http.Handler) http.Handler {
+func NewContext(debugResponses bool, mc ring.MemcacheRing, log srv.LowLevelLogger, proxyClientFactory client.ProxyClient) func(http.Handler) http.Handler {
 	return func(next http.Handler) http.Handler {
 		return &ProxyContextMiddleware{
-			Cache:             mc,
-			log:               log,
-			next:              next,
-			proxyDirectClient: proxyDirectClient,
-			debugResponses:    debugResponses,
+			Cache:              mc,
+			log:                log,
+			next:               next,
+			proxyClientFactory: proxyClientFactory,
+			debugResponses:     debugResponses,
 		}
 	}
 }
