+++ proxyserver/middleware/tempurl_test.go
@@ -26,6 +26,9 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/troubling/hummingbird/client"
 	"github.com/troubling/hummingbird/common"
+	"github.com/troubling/hummingbird/common/conf"
+	"github.com/troubling/hummingbird/common/srv"
+	"github.com/troubling/hummingbird/common/test"
 	"go.uber.org/zap"
 )
 
@@ -195,8 +198,11 @@ func TestTempurlMiddleware400PuttingManifest(t *testing.T) {
 
 func TestTempurlMiddleware401NoKeys(t *testing.T) {
 	r := httptest.NewRequest("PUT", "/v1/a/c/o?temp_url_sig=ABCDEF&temp_url_expires=9999999999", nil)
+	f, err := client.NewProxyClient(staticPolicyList, srv.NewTestConfigLoader(&test.FakeRing{}),
+		nil, "", "", "", "", "", conf.Config{})
+	require.Nil(t, err)
 	ctx := &ProxyContext{
-		C: client.NewProxyClient(&client.ProxyDirectClient{}, nil, map[string]*client.ContainerInfo{
+		C: f.NewRequestClient(nil, map[string]*client.ContainerInfo{
 			"container/a/c": {Metadata: map[string]string{}},
 		}, zap.NewNop()),
 		accountInfoCache: map[string]*AccountInfo{
@@ -213,8 +219,11 @@ func TestTempurlMiddleware401NoKeys(t *testing.T) {
 
 func TestTempurlMiddleware401WrongKeys(t *testing.T) {
 	r := httptest.NewRequest("PUT", "/v1/a/c/o?temp_url_sig=ABCDEF&temp_url_expires=9999999999", nil)
+	f, err := client.NewProxyClient(staticPolicyList, srv.NewTestConfigLoader(&test.FakeRing{}),
+		nil, "", "", "", "", "", conf.Config{})
+	require.Nil(t, err)
 	ctx := &ProxyContext{
-		C: client.NewProxyClient(&client.ProxyDirectClient{}, nil, map[string]*client.ContainerInfo{
+		C: f.NewRequestClient(nil, map[string]*client.ContainerInfo{
 			"container/a/c": {Metadata: map[string]string{"Temp-Url-Key": "ABCD", "Temp-Url-Key-2": "012345"}},
 		}, zap.NewNop()),
 		accountInfoCache: map[string]*AccountInfo{
@@ -232,8 +241,11 @@ func TestTempurlMiddleware401WrongKeys(t *testing.T) {
 func TestTempurlMiddlewareContainerKey(t *testing.T) {
 	r := httptest.NewRequest("GET", "/v1/a/c/o?temp_url_sig=f2d61be897a27c03ac9a0dac3a8c4f6ce3a3d623&"+
 		"temp_url_expires=9999999999", nil)
+	f, err := client.NewProxyClient(staticPolicyList, srv.NewTestConfigLoader(&test.FakeRing{}),
+		nil, "", "", "", "", "", conf.Config{})
+	require.Nil(t, err)
 	ctx := &ProxyContext{
-		C: client.NewProxyClient(&client.ProxyDirectClient{}, nil, map[string]*client.ContainerInfo{
+		C: f.NewRequestClient(nil, map[string]*client.ContainerInfo{
 			"container/a/c": {Metadata: map[string]string{"Temp-Url-Key": "mykey"}},
 		}, zap.NewNop()),
 		accountInfoCache: map[string]*AccountInfo{"account/a": {Metadata: map[string]string{}}},
@@ -261,8 +273,11 @@ func TestTempurlMiddlewareContainerKey(t *testing.T) {
 func TestTempurlMiddlewarePath(t *testing.T) {
 	r := httptest.NewRequest("GET", "/v1/a/c/o123?temp_url_sig=058e0771c69f7e1eb1eacbd68396920fd06ff261&"+
 		"temp_url_expires=9999999999&temp_url_prefix=o", nil)
+	f, err := client.NewProxyClient(staticPolicyList, srv.NewTestConfigLoader(&test.FakeRing{}),
+		nil, "", "", "", "", "", conf.Config{})
+	require.Nil(t, err)
 	ctx := &ProxyContext{
-		C: client.NewProxyClient(&client.ProxyDirectClient{}, nil, map[string]*client.ContainerInfo{
+		C: f.NewRequestClient(nil, map[string]*client.ContainerInfo{
 			"container/a/c": {Metadata: map[string]string{"Temp-Url-Key": "mykey"}},
 		}, zap.NewNop()),
 		accountInfoCache: map[string]*AccountInfo{"account/a": {Metadata: map[string]string{}}},
@@ -284,8 +299,11 @@ func TestTempurlMiddlewarePath(t *testing.T) {
 func TestTempurlMiddlewareAccountKey(t *testing.T) {
 	r := httptest.NewRequest("GET", "/v1/a/c/o?temp_url_sig=f2d61be897a27c03ac9a0dac3a8c4f6ce3a3d623&"+
 		"temp_url_expires=9999999999", nil)
+	f, err := client.NewProxyClient(staticPolicyList, srv.NewTestConfigLoader(&test.FakeRing{}),
+		nil, "", "", "", "", "", conf.Config{})
+	require.Nil(t, err)
 	ctx := &ProxyContext{
-		C: client.NewProxyClient(&client.ProxyDirectClient{}, nil, map[string]*client.ContainerInfo{
+		C: f.NewRequestClient(nil, map[string]*client.ContainerInfo{
 			"container/a/c": {Metadata: map[string]string{}},
 		}, zap.NewNop()),
 		accountInfoCache: map[string]*AccountInfo{
